Logseq file path encode, Auto source helpers rename
