+++ src/promnesia/sources/auto_logseq.py
@@ -1,4 +1,5 @@
 import os.path
+import urllib.parse
 
 def logseq_replacer(path: str, root: str) -> str:
     if not path.startswith("editor://") or not (path.endswith('.md') or path.endswith('.org')):
@@ -6,7 +7,8 @@ def logseq_replacer(path: str, root: str) -> str:
         
     graph = os.path.basename(root)
     page_name = os.path.basename(path).rsplit('.', 1)[0]
+    encoded_page_name = urllib.parse.quote(page_name)
     
-    uri = f"logseq://graph/{graph}?page={page_name}"
+    uri = f"logseq://graph/{graph}?page={encoded_page_name}"
 
     return uri
