+++ src/promnesia/sources/auto_obsidian.py
